+++ packages/client/src/hooks/useAuth.ts
@@ -1,6 +1,6 @@
 import { useAppSelector } from '@/hooks/redux'
 
-export default function useAuth(): boolean  {
+export default function useAuth(): boolean | null  {
   const { isLoggedIn } = useAppSelector(state => state.auth)
 
   return isLoggedIn
