+++ packages/client/src/router.tsx
@@ -1,4 +1,4 @@
-import { Routes, Route } from 'react-router-dom'
+import { Routes, Route, Navigate } from 'react-router-dom'
 import Offline from '@/pages/offline'
 import SignIn from '@/pages/sign-in'
 import SignUp from '@/pages/sign-up'
@@ -9,6 +9,9 @@ import Error404 from '@/pages/error404'
 import Error500 from '@/pages/error500'
 import Forum from '@/pages/forum'
 import Profile from '@/pages/profile'
+import ProtectRoute from '@/components/protect-route'
+import ProtectRouteForSignIn from '@/components/protect-route-for-sign-in'
+import Comments from '@/pages/comments'
 
 export enum RoutePaths {
   SIGNIN = '/sign-in',
@@ -18,23 +21,34 @@ export enum RoutePaths {
   ERROR404 = '/404',
   ERROR500 = '/500',
   FORUM = '/forum',
-  PROFILE = '/profile'
+  PROFILE = '/profile',
+  COMMENTS = '/forum/:title',
 }
 
 export const Router = () => {
   return (
     <Routes>
-      <Route path='/offline' element={<Offline />} />
-      <Route path={RoutePaths.SIGNIN} element={<SignIn />} />
-      <Route path={RoutePaths.SIGNUP} element={<SignUp />} />
-      <Route element={<FullScreen/>}>
-        <Route path={RoutePaths.GAME} element={<Game />} />
+      <Route
+        element={<ProtectRouteForSignIn redirectTo={RoutePaths.GAME} />}
+        loader={async () => Promise.resolve(true)}>
+        <Route path={RoutePaths.SIGNIN} element={<SignIn />} />
+        <Route path={RoutePaths.SIGNUP} element={<SignUp />} />
       </Route>
-      <Route path={RoutePaths.LEADERBOARD} element={<Leaderboard />} />
+      <Route element={<ProtectRoute redirectTo={RoutePaths.SIGNIN} />}>
+        <Route element={<Leaderboard />} path={RoutePaths.LEADERBOARD} />
+        <Route element={<FullScreen />}>
+          <Route path={RoutePaths.GAME} element={<Game />} />
+        </Route>
+        <Route path={RoutePaths.FORUM} element={<Forum />} />
+        <Route element={<Forum />} path={RoutePaths.FORUM} />
+        <Route path={RoutePaths.COMMENTS} element={<Comments />} />
+        <Route element={<Profile />} path={RoutePaths.PROFILE} />
+        <Route path={RoutePaths.PROFILE} element={<Profile />} />
+      </Route>
+      <Route path='/offline' element={<Offline />} />
       <Route path={RoutePaths.ERROR404} element={<Error404 />} />
       <Route path={RoutePaths.ERROR500} element={<Error500 />} />
-      <Route path={RoutePaths.FORUM} element={<Forum />} />
-      <Route path={RoutePaths.PROFILE} element={<Profile />} />
+      <Route path='*' element={<Navigate to={RoutePaths.ERROR404} replace />} />
     </Routes>
   )
 }
