+++ packages/client/vite.config.ts
@@ -51,9 +51,9 @@ export default defineConfig(({ command, mode }) => {
       rollupOptions: {
         //TODO: add hash to filename after service worker fix dynamic names
         output: {
-          assetFileNames: 'assets/[name][extname]',
-          chunkFileNames: '[name].js',
-          entryFileNames: '[name].js'
+          assetFileNames: 'assets/[name][hash][extname]',
+          chunkFileNames: '[name][hash].js',
+          entryFileNames: '[name][hash].js'
         }
       }
     }
